+++ OSM_Unity_Client/Assets/Scripts/LandmarkHeuristic.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public class LandmarkHeuristic : AStarHeuristic
 {
-    private readonly Dictionary<long, long> distances;
+    private readonly Dictionary<long, float> distances;
     private bool behind;
 
     public LandmarkHeuristic(Landmark landmark, bool behind)
